% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bands}
\alias{bands}
\title{Venus bands}
\format{
A \code{data.frame} with 15 rows (bands) and fourt columns: \code{band} (band ID, from 1 to 12), \code{lower} (lower wavelength, in micrometers), \code{upper} (lower wavelength, in micrometers), and \code{esun} (solar irradiance, in W/m^2).
}
\usage{
bands
}
\description{
A \code{data.frame} with information about Venus satellite bands
}
\keyword{datasets}
