% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_venus_metadata.R
\name{get_venus_metadata}
\alias{get_venus_metadata}
\title{Extract Venus satellite image metadata}
\usage{
get_venus_metadata(x, band)
}
\arguments{
\item{x}{A \code{list} with the Venus image metadata, as returned by reading the Venus \code{HDR} file with \code{readLines} and parsing to \code{list} with \code{XML::xmlToList}}

\item{band}{The Venus band for which to return meatadata, \code{numeric} of length 1 between \code{1} and \code{12}}
}
\value{
\code{list} with the Venus image metadata values
elevation in km...
}
\description{
Extracts Venus satellite image metadata, as required for atrmospheric correction using \code{sixs_params}, from a Venus \code{HDR} metadata file.
}
\examples{
library(XML)
data(venus1m)
venus1m = xmlToList(venus1m)
get_venus_metadata(venus1m, 3)
}
