% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad_to_refl.R
\name{rad_to_refl}
\alias{rad_to_refl}
\title{Convert a TOA radiance image to a TOA reflectance image}
\usage{
rad_to_refl(toa_rad, date, esun, solar_zenith_angle)
}
\arguments{
\item{toa_rad}{A \code{stars} image with top of atmosphere radiance values}

\item{date}{Acquisition date, needs to be \code{Date} of length 1}

\item{esun}{External solar radiation [W/m^2]. For Venus satellite bands, \code{esun} values are provided in the \code{bands} dataset.}

\item{solar_zenith_angle}{Solar zenith angle [degrees, between -180 and 180]}
}
\value{
A \code{stars} image with top of atmosphere reflectance values
}
\description{
Converts a TOA radiance image to a TOA reflectance image.
}
\details{
Top of atmosphere reflectance can be caulculated using the following formula:
 
TOA = (pi * L * d^2) / (ESUN * cos(solar_zenith_angle))

where 'TOA' is the top of atmosphere reflectance, 'L' is the top of atmosphere radiance, 'ESUN' is the external solar radiation and 'solar_zenith_angle' is the solar zenith angle.
}
\examples{

library(stars)
library(XML)

data(venus1)
data(venus1m)

# Get metadata
l = xmlToList(venus1m)
m = get_venus_metadata(l, band = 1)

# TOA radiance -> TOA reflectance
toa_rad =  venus1 * 0.001
toa_refl = NULL
for(i in 1:dim(venus1)[3]) {
    toa_refl[[i]] = rad_to_refl(
       toa_rad = toa_rad[,,,i,drop=TRUE],
       date = m$date,
       esun = bands$esun[i],
       solar_zenith_angle = m$solar_zenith_angle
    )
}
toa_refl = do.call(c, toa_refl)
toa_refl = st_redimension(toa_refl)

# Plot rasters
plot(toa_rad)
plot(toa_refl)

# Profile for one pixel
toa_rad1 = venus1[[1]][100,100,]
toa_refl1 = toa_refl[[1]][100,100,]

# Plot reflectance profile
w = rowMeans(bands[c("lower", "upper")])
plot(
    w, toa_rad1, type = "l", 
    xlab = "Wavelength (micrometers)", ylab = "TOA radiance"
)
plot(
    w, toa_refl1, type = "l", 
    xlab = "Wavelength (micrometers)", ylab = #' "TOA reflectance"
)
}
