% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refl_to_rad.R
\name{refl_to_rad}
\alias{refl_to_rad}
\title{Convert a TOA reflectance image to a TOA radiance image}
\usage{
refl_to_rad(toa_refl, date, esun, solar_zenith_angle)
}
\arguments{
\item{toa_refl}{A \code{stars} image with top of atmosphere reflectance values}

\item{date}{Acquisition date, needs to be \code{Date} of length 1}

\item{esun}{External solar radiation [W/m^2]. For Venus satellite bands, \code{esun} values are provided in the \code{bands} dataset.}

\item{solar_zenith_angle}{Solar zenith angle [degrees, between -180 and 180]}
}
\value{
A \code{stars} image with top of atmosphere reflectance values
}
\description{
Converts a TOA reflectance image to a TOA radiance image.
}
\details{
Top of atmosphere radiance can be caulculated using the following formula:
 
L = TOA * (ESUN * cos(solar_zenith_angle) / (pi * d^2)

where 'L' is the top of atmosphere radiance, 'TOA' is the top of atmosphere reflectance, 'ESUN' is the external solar radiation and 'solar_zenith_angle' is the solar zenith angle.
}
\examples{
# Band 3 in sample image 'venus1'
library(stars)
library(XML)

# Reflectance Venus image, band 3
refl = venus1[,,,3,drop=TRUE] / 1000

# Metadata
l = xmlToList(venus1m)
m = get_venus_metadata(l, band = 1)

# Specific pixel value: reflectance
refl[[1]][100,100]  

# Conversion from reflectance to radiance
rad = refl_to_rad(
    toa_refl = refl,
    date = m$date,
    esun = bands$esun[3],
    solar_zenith_angle = m$solar_zenith
)

# Specific pixel value: radiance
rad[[1]][100,100]
}
