% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toa_to_boa.R
\name{toa_to_boa}
\alias{toa_to_boa}
\title{Convert TOA radiance image to a BOA reflectance image}
\usage{
toa_to_boa(toa_rad, params)
}
\arguments{
\item{toa_rad}{A \code{stars} image with top of atmosphere radiance values}

\item{params}{A \code{numeric} vector of length 3 with the atmospheric correction coefficients (Xa, Xb, Xc, in that order), as returned by \code{sixs_params}}
}
\value{
A \code{stars} image with bottom of atmosphere reflectance values
}
\description{
Converts a TOA radiance image to a BOA reflectance image.
}
\details{
The atmospheric correction coefficients (Xa, Xb, Xc) returned by the \code{sixs_params} function can be used to calculate bottom of the atmosphere (BOA) reflectance from top of the atmosphere (TOA) radiance using the following formulas (1) and (2) (Zhao, Tamura, and Takahashi 2001):

(1) y = Xa * TOA - Xb

(2) BOA = y / (1 + Xc * y)

where 'y' is an intermediate variable, 'TOA' is the top of the atmosphere radiance, 'BOA' is the bottom of the atmosphere reflectance, and 'Xa', 'Xb', 'Xc' are the correction coefficents returned by the \code{sixs_params} function.
}
\examples{


}
\references{
Zhao, W., Tamura, M., & Takahashi, H. (2001). Atmospheric and spectral corrections for estimating surface albedo from satellite data using 6S code. Remote sensing of Environment, 76(2), 202-212.
}
