% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{venus1}
\alias{venus1}
\title{Sample Venus image 1: Raster}
\format{
A \code{stars} raster, where values represent TOA reflectance multiplied by 1000.
}
\usage{
venus1
}
\description{
A three-dimensional \code{stars} raster, with 12 layers (spectral bands), as resulting from reading a Venus satellite image \code{TIF} file with:
}
\details{
\code{read_stars("VE_VM01_VSC_PDTIMG_L1VALD_ISRAEL03_20190204.DBL.TIF_crop.TIF")}

then cropping a small (600*600 m^2) area and selecting just the first 12 image bands.
}
\keyword{datasets}
