% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{venus1_boa}
\alias{venus1_boa}
\title{Sample Venus image 1: BOA reflectance}
\format{
A \code{stars} raster
}
\usage{
venus1_boa
}
\description{
A three-dimensional \code{stars} raster, with 12 layers (spectral bands), after converting \code{venus1} from TOA reflectance to BOA reflectance.
}
\keyword{datasets}
