% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{venus1m}
\alias{venus1m}
\title{Sample Venus image 1: Metadata}
\format{
A \code{character} vector of length 1
}
\usage{
venus1m
}
\description{
A \code{character} vector of length 1, with the contents of the XML document with image metadata, as resulting from reading a Venus satellite image metadata \code{HDR} file with:
}
\details{
\code{paste0(readLines("VE_VM01_VSC_L1VALD_ISRAEL03_20190204.HDR"), collapse = "")}
}
\keyword{datasets}
