% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{venus2_boa}
\alias{venus2_boa}
\title{Sample Venus image 2: BOA reflectance}
\format{
A \code{stars} raster
}
\usage{
venus2_boa
}
\description{
A three-dimensional \code{stars} raster, with 12 layers (spectral bands), after converting \code{venus2} from TOA reflectance to BOA reflectance.
}
\keyword{datasets}
