% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{venus_validation}
\alias{venus_validation}
\title{Venus validation}
\format{
A \code{data.frame}
}
\usage{
venus_validation
}
\description{
A \code{data.frame} with metadata for the two sample Venus satellite images, extracted from (\code{venus1m} and \code{venus2m}). Provided for validation purposes.
}
\keyword{datasets}
