% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad_to_refl.R
\name{rad_to_refl}
\alias{rad_to_refl}
\title{Convert a TOA radiance image to a TOA reflectance image}
\usage{
rad_to_refl(toa_rad, date, esun, solar_zenith_angle)
}
\arguments{
\item{toa_rad}{A \code{stars} image with top of atmosphere radiance values}

\item{date}{Acquisition date, needs to be \code{Date} of length 1}

\item{esun}{External solar radiation [W/m^2]. For Venus satellite bands, \code{esun} values are provided in the \code{bands} dataset.}

\item{solar_zenith_angle}{Solar zenith angle [degrees, between -180 and 180]}
}
\value{
A \code{stars} image with top of atmosphere reflectance values
}
\description{
Converts a TOA radiance image to a TOA reflectance image.
}
\details{
Top of atmosphere reflectance can be caulculated using the following formula:
 
TOA = (pi * L * d^2) / (ESUN * cos(solar_zenith_angle))

where 'TOA' is the top of atmosphere reflectance, 'L' is the top of atmosphere radiance, 'ESUN' is the external solar radiation and 'solar_zenith_angle' is the solar zenith angle.
}
\examples{

}
